+++ VCG_Raytracer/src/utils/Vec2.java
@@ -2,10 +2,10 @@ package utils;
 
 
 public class Vec2 {
-    public int x;
-    public int y;
+    public float x;
+    public float y;
 
-    public Vec2(int x, int y){
+    public Vec2(float x, float y){
         this.x = x;
         this.y = y;
     }
